pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.1.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "schedule>=1.2.2",
    "tweepy>=4.16.0",
    "vadersentiment>=3.3.2",
]
